+++ src/rich_click/decorators.py
@@ -1,3 +1,4 @@
+import sys
 from gettext import gettext
 from typing import TYPE_CHECKING, Any, Callable, Dict, Mapping, Optional, Type, TypeVar, Union, cast, overload
 
@@ -5,14 +6,18 @@ from click import Command, Context, Group, Parameter
 from click import command as click_command
 from click import option as click_option
 from click import pass_context as click_pass_context
-from typing_extensions import Concatenate, ParamSpec
 
-from rich_click._compat_click import CLICK_IS_BEFORE_VERSION_8X
-from rich_click.rich_command import RichCommand, RichGroup, RichMultiCommand  # noqa: F401
+from rich_click.rich_command import RichCommand, RichGroup
 from rich_click.rich_context import RichContext
 from rich_click.rich_help_configuration import RichHelpConfiguration
 
 from . import rich_click  # noqa: F401
+
+
+if sys.version_info < (3, 10):
+    from typing_extensions import Concatenate, ParamSpec
+else:
+    from typing import Concatenate, ParamSpec
 
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -163,21 +168,6 @@ def rich_config(
         )
         console = help_config
 
-    if CLICK_IS_BEFORE_VERSION_8X:
-
-        def decorator_with_warning(obj: FC) -> FC:
-            import warnings
-
-            warnings.warn(
-                "`rich_config()` does not work with versions of click prior to version 8.0.0."
-                " Please update to a newer version of click to use this functionality.",
-                RuntimeWarning,
-                stacklevel=2,
-            )
-            return obj
-
-        return decorator_with_warning
-
     def decorator(obj: FC) -> FC:
         extra: Dict[str, Any] = {}
         if console is not None:
@@ -233,7 +223,7 @@ def help_option(*param_decls: str, **kwargs: Any) -> Callable[[FC], FC]:
         if value and not ctx.resilient_parsing:
             # Avoid click.echo() because it ignores console settings like force_terminal.
             # Also, do not print() if empty string; assume console was record=False.
-            print(ctx.get_help(), end="")
+            print(ctx.get_help())
             ctx.exit()
 
     if not param_decls: