+++ tests/fixtures/defaults.py
@@ -1,28 +1,17 @@
 import rich_click as click
 
 
-# click.rich_click.SHOW_ARGUMENTS = True
-# click.rich_click.GROUP_ARGUMENTS_OPTIONS = True
-
-
-@click.command()
-@click.argument("input", type=click.Path(), required=True)
-@click.argument("output", type=click.Path())
-@click.option("--type", default="files", show_default=True, help="Type of file to sync")
+@click.group()
 @click.option("--debug", "-d", is_flag=True, help="Enable debug mode")
 @click.option(
     "--environment",
     "-e",
     type=click.Choice(["dev", "staging", "prod"]),
-    show_default="current",
     envvar="MY_ENV",
     show_envvar=True,
     help="Sync to what environment",
 )
 def cli(
-    input: str,
-    output: str,
-    type: str,
     debug: bool,
     environment: str,
 ) -> None:
@@ -35,11 +24,15 @@ def cli(
     You can try using --help at the top level and also for
     specific group subcommands.
     """
-    print(f"Input: {input}")
-    print(f"Output: {output}")
     print(f"Environment: {environment}")
     print(f"Debug mode is {'on' if debug else 'off'}")
-    print(f"Syncing files of type {type}")
+
+
+@cli.command()
+@click.option("--files", default="*", show_default="All files", help="What files to download")
+def download(files: str) -> None:
+    """Download files"""
+    print(f"Downloading {files}")
 
 
 if __name__ == "__main__":